Prepared code for smarted evaluation function:  - implemented move-ordering by evaluation of successors with --mo=2    (this seems to work best *without* killer heuristic)  - changed value type to float  - written more complex evaluation function (currently weighted to match old    player)

+++ Eval.h
@@ -4,8 +4,8 @@
 #include "Game.h"
 
 /* Type of values used to evaluate game positions: */
-#define VAL_FMT "%d"
-typedef signed short val_t;
+#define VAL_FMT "%f"
+typedef float val_t;
 extern const val_t min_val, max_val;
 
 /* Board evaluation functions, used by AI: */
